+++ boot/uefi/core/src/ffi/system.rs
@@ -30,16 +30,16 @@ pub struct Table {
 
 impl Table {
     pub const SIGNATURE : u64 = 0x5453595320494249;
-    pub const LATEST_REVISION : u64 = Self::REVISION_2_70;
-    pub const REVISION_2_70 : u64 = ((2<<16) | (70));
-    pub const REVISION_2_60 : u64 = ((2<<16) | (60));
-    pub const REVISION_2_50 : u64 = ((2<<16) | (50));
-    pub const REVISION_2_40 : u64 = ((2<<16) | (40));
-    pub const REVISION_2_31 : u64 = ((2<<16) | (31));
-    pub const REVISION_2_30 : u64 = ((2<<16) | (30));
-    pub const REVISION_2_20 : u64 = ((2<<16) | (20));
-    pub const REVISION_2_10 : u64 = ((2<<16) | (10));
-    pub const REVISION_2_00 : u64 = ((2<<16) | (00));
-    pub const REVISION_1_10 : u64 = ((1<<16) | (10));
-    pub const REVISION_1_02 : u64 = ((1<<16) | (02));
+    pub const LATEST_REVISION : u32 = Self::REVISION_2_70;
+    pub const REVISION_2_70 : u32 = ((2<<16) | (70));
+    pub const REVISION_2_60 : u32 = ((2<<16) | (60));
+    pub const REVISION_2_50 : u32 = ((2<<16) | (50));
+    pub const REVISION_2_40 : u32 = ((2<<16) | (40));
+    pub const REVISION_2_31 : u32 = ((2<<16) | (31));
+    pub const REVISION_2_30 : u32 = ((2<<16) | (30));
+    pub const REVISION_2_20 : u32 = ((2<<16) | (20));
+    pub const REVISION_2_10 : u32 = ((2<<16) | (10));
+    pub const REVISION_2_00 : u32 = ((2<<16) | (00));
+    pub const REVISION_1_10 : u32 = ((1<<16) | (10));
+    pub const REVISION_1_02 : u32 = ((1<<16) | (02));
 }
